refactor (#8): 리뷰에따라 수정
