+++ src/main/java/io/toyproject/bookjeokclone/model/dto/BoardRequest.java
@@ -1,9 +1,11 @@
 package io.toyproject.bookjeokclone.model.dto;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 
 
 @Data
+@AllArgsConstructor
 public class BoardRequest {
