+++ src/main/java/io/toyproject/bookjeokclone/service/BoardService.java
@@ -9,7 +9,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-@Transactional
 public class BoardService {
 
     private final BoardRepository boardRepository;
