+++ src/main/resources/templates/board/write.html
@@ -6,7 +6,7 @@
 </head>
 <body>
     <div class="container">
-        <form action="/board/write" method="post">
+        <form action="/boards" method="post">
             <div class="form-group">
                 <label for="writer">작성자</label>
                 <input type="text" id="writer" name="writer" placeholder="작성자를 입력하세요">
